Revert "Revert "Slambook 2014 first remote commit""
This reverts commit 90aadca8bdfe14b9fa2d4f280f46de1f50975974.

+++ src/java/com/slambook/services/interfaces/ImageLikeServiceInt.java
@@ -13,11 +13,11 @@ import java.util.ArrayList;
  */
 public interface ImageLikeServiceInt {
    
-     public int addLike(ImageLikes imageLikes);
-     public int removeLike(ImageLikes imageLikes);
-     public ArrayList getImageLikes(int imageId);
-     public int getImageLikesCount(int imageId);
-     public int returnImageLikesCount(int imageId);
-     public boolean isAlreadyLiked(int imageId,int userId);
+     public long addLike(ImageLikes imageLikes);
+     public long removeLike(ImageLikes imageLikes);
+     public ArrayList getImageLikes(long imageId);
+     public long getImageLikesCount(long imageId);
+     public long returnImageLikesCount(long imageId);
+     public boolean isAlreadyLiked(long imageId,long userId);
    
 }
